requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0

# Search Engines
pysolr==3.9.0
chromadb==0.4.22

# Embeddings
sentence-transformers>=2.3.1
torch>=2.0.0

# Data Processing - NumPy pinned to <2.0 for ChromaDB 0.4.22 compatibility
pandas>=2.1.0
numpy>=1.24.0,<2.0.0

# Utilities
pyyaml>=6.0.0
python-dateutil>=2.8.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0

# Logging
colorlog>=6.8.0
